requirements.txt
Django==3.0
django-autocomplete-light==3.5.0
django-extensions==2.2.5
factory-boy==2.12.0
psycopg2==2.8.3
requests==2.22.0
responses==0.10.9